# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pathwise CC API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles/{profileId}/generate-matches:
    post:
      operationId: generateMatches
      tags: [matches]
      summary: Score universities for profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ranked university matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UniversityMatch"
  /profiles/{profileId}/matches:
    get:
      operationId: getMatches
      tags: [matches]
      summary: Get cached matches for profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cached university matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UniversityMatch"
...
  /pathways/{pathwayId}/generate-guidebook:
    post:
      operationId: generateGuidebook
      tags: [guidebooks]
      summary: Generate AI guidebook for pathway
      parameters:
        - name: pathwayId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated guidebook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guidebook"
        "429":
          description: Rate limit exceeded
  /guidebooks/{guidebookId}:
    get:
      operationId: getGuidebook
      tags: [guidebooks]
      summary: Get guidebook
      parameters:
        - name: guidebookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Guidebook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guidebook"
    delete:
      operationId: deleteGuidebook
      tags: [guidebooks]
      summary: Delete guidebook
      parameters:
        - name: guidebookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /profiles/{profileId}/guidebooks:
    get:
      operationId: getGuidebooks
      tags: [guidebooks]
      summary: List guidebooks for profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of guidebooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guidebook"
...
  /profiles/{profileId}/recommended-scholarships:
    post:
      operationId: getRecommendedScholarships
      tags: [opportunities]
      summary: Get filtered scholarship recommendations
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recommended scholarships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scholarship"
  /profiles/{profileId}/recommended-opportunities:
    post:
      operationId: getRecommendedOpportunities
      tags: [opportunities]
      summary: Get filtered opportunity recommendations
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recommended opportunities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Opportunity"
...
components:
  schemas:
...
    UniversityMatch:
      type: object
      properties:
        universityId:
          type: string
        name:
          type: string
        system:
          type: string
        location:
          type: string
        compatibilityScore:
          type: integer
        fitLabel:
          type: string
        fitColor:
          type: string
        gpaRangeMin:
          type: number
        gpaRangeRecommended:
          type: number
        costCategory:
          type: string
        transferFriendliness:
          type: integer
        officialTransferUrl:
          type: string
        notes:
          type: string
        breakdown:
          $ref: "#/components/schemas/ScoreBreakdown"
        majors:
          type: array
          items:
            type: string
        careerTags:
          type: array
          items:
            type: string
      required: [universityId, name, compatibilityScore, fitLabel]
    ScoreBreakdown:
      type: object
      properties:
        majorAlignment:
          type: number
        gpaFit:
          type: number
        coursework:
          type: number
        costFit:
          type: number
        locationFit:
          type: number
        careerFit:
          type: number
        experienceAccess:
          type: number
        timeFit:
          type: number
        supportFit:
          type: number
      required: [majorAlignment, gpaFit, costFit, locationFit, careerFit]
...
    Guidebook:
      type: object
      properties:
        id:
          type: integer
        pathwayId:
          type: integer
        profileId:
          type: integer
        contentMarkdown:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, pathwayId, profileId]
    Scholarship:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        amount:
          type: string
        deadline:
          type: string
        minGpa:
          type: number
          nullable: true
        eligibilityTags:
          type: array
          items:
            type: string
        description:
          type: string
        sourceUrl:
          type: string
        sourceDate:
          type: string
        isEstimate:
          type: boolean
      required: [id, name, amount]
    Opportunity:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        description:
          type: string
        eligibilityTags:
          type: array
          items:
            type: string
        applicationUrl:
          type: string
        sourceDate:
          type: string
        isEstimate:
          type: boolean
      required: [id, name, type]
...