# lib/api-zod/src/generated/types/guidebook.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pathwise CC API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Guidebook {
  id: number;
  pathwayId: number;
  profileId: number;
  contentMarkdown?: string | null;
  title?: string | null;
  createdAt?: string;
  updatedAt?: string;
}