# lib/api-zod/src/generated/types/opportunity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pathwise CC API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Opportunity {
  id: string;
  name: string;
  type: string;
  description?: string;
  eligibilityTags?: string[];
  applicationUrl?: string;
  sourceDate?: string;
  isEstimate?: boolean;
}