# lib/api-zod/src/generated/types/scoreBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pathwise CC API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreBreakdown {
  majorAlignment: number;
  gpaFit: number;
  coursework?: number;
  costFit: number;
  locationFit: number;
  careerFit: number;
  experienceAccess?: number;
  timeFit?: number;
  supportFit?: number;
}